requirements.txt
langchain-groq
colorama>=0.4.6
python-dotenv>=1.0.0
langchain